feat(frontend): add trace headers to API requests for improved logging
- Enhanced various API functions to accept optional trace headers, allowing for better observability and debugging.
- Updated the `sendRestGqlRequest` utility to incorporate trace headers into requests.
- Implemented structured logging using the @kids-reporter/logger package across multiple components, replacing console logging for consistency.
- Improved error handling to include trace context in log emissions, facilitating easier tracking of issues.
