+++ packages/frontend/src/api/call-baodaozai-intro.ts
@@ -6,12 +6,14 @@ import {
 import { sendRestGqlRequest } from '@/utils/send-rest-gql'
 
 export async function getCallBaodaozaiIntroContent(
-  variables: GetCallBaodaozaiIntroQueryVariables
+  variables: GetCallBaodaozaiIntroQueryVariables,
+  traceHeaders?: Record<string, string>
 ): Promise<string | undefined> {
   const data = await sendRestGqlRequest<GetCallBaodaozaiIntroQuery>({
     operation: 'call-baodaozai-intro',
     method: 'GET',
     variables,
+    traceHeaders,
   })
 
   return data?.data?.data?.callBaodaozaiIntro?.content ?? undefined
