+++ packages/frontend/src/api/popular-keywords.ts
@@ -6,12 +6,14 @@ import {
 import { sendRestGqlRequest } from '@/utils/send-rest-gql'
 
 export const getPopularKeywords = async (
-  variables?: GetPopularKeywordsQueryVariables
+  variables?: GetPopularKeywordsQueryVariables,
+  traceHeaders?: Record<string, string>
 ) => {
   const response = await sendRestGqlRequest<GetPopularKeywordsQuery>({
     operation: 'popular-keywords',
     method: 'GET',
     variables: variables ?? {},
+    traceHeaders,
   })
   return response?.data?.data?.popularKeywords
 }
