+++ packages/frontend/src/api/post.ts
@@ -13,44 +13,54 @@ import {
 import { sendRestGqlRequest } from '@/utils/send-rest-gql'
 
 export const getLatestPosts = async (
-  variables: GetLatestPostsQueryVariables
+  variables: GetLatestPostsQueryVariables,
+  traceHeaders?: Record<string, string>
 ) => {
   const response = await sendRestGqlRequest<GetLatestPostsQuery>({
     operation: 'latest-posts',
     method: 'GET',
     variables,
+    traceHeaders,
   })
   return response?.data?.data?.posts
 }
 
-export const getPost = async (variables: GetPostQueryVariables) => {
+export const getPost = async (
+  variables: GetPostQueryVariables,
+  traceHeaders?: Record<string, string>
+) => {
   const response = await sendRestGqlRequest<GetPostQuery>({
     operation: 'post-detail',
     method: 'GET',
     variables,
+    traceHeaders,
   })
   return response?.data?.data?.post
 }
 
 export const getPostMeta = async (
-  variables: GetPostMetaQueryVariables
+  variables: GetPostMetaQueryVariables,
+  traceHeaders?: Record<string, string>
 ): Promise<GetPostMetaQuery['post']> => {
   const response = await sendRestGqlRequest<GetPostMetaQuery>({
     operation: 'post-meta',
     method: 'GET',
     variables,
+    traceHeaders,
   })
   return response?.data?.data?.post
 }
 
 export const getPostsEssayAnswersWithLikes = async (
-  variables: GetPostsEssayAnswersWithLikesQueryVariables
+  variables: GetPostsEssayAnswersWithLikesQueryVariables,
+  traceHeaders?: Record<string, string>
 ) => {
   const response = await sendRestGqlRequest<GetPostsEssayAnswersWithLikesQuery>(
     {
       operation: 'posts-essay-answers-with-likes',
       method: 'GET',
       variables,
+      traceHeaders,
     }
   )
   return response?.data?.data?.posts
@@ -58,13 +68,16 @@ export const getPostsEssayAnswersWithLikes = async (
 
 export const getPostEssayQuestionsByPostSlug = async ({
   slug,
+  traceHeaders,
 }: {
   slug: string
+  traceHeaders?: Record<string, string>
 }) => {
   const response = await sendRestGqlRequest<GetPostEssayQuestionsQuery>({
     operation: 'post-essay-questions',
     method: 'GET',
     variables: { where: { slug } },
+    traceHeaders,
   })
   return response?.data?.data?.post
 }
