+++ packages/frontend/src/api/project.ts
@@ -8,24 +8,28 @@ import {
 import { sendRestGqlRequest } from '@/utils/send-rest-gql'
 
 export const getTopicProjects = async (
-  variables: GetTopicProjectsQueryVariables
+  variables: GetTopicProjectsQueryVariables,
+  traceHeaders?: Record<string, string>
 ) => {
   const response = await sendRestGqlRequest<GetTopicProjectsQuery>({
     operation: 'topic-projects',
     method: 'GET',
     variables,
+    traceHeaders,
   })
 
   return response?.data?.data?.projects
 }
 
 export const getTopicProjectsPaged = async (
-  variables: GetProjectsQueryVariables
+  variables: GetProjectsQueryVariables,
+  traceHeaders?: Record<string, string>
 ) => {
   const response = await sendRestGqlRequest<GetProjectsQuery>({
     operation: 'projects-paged',
     method: 'GET',
     variables,
+    traceHeaders,
   })
 
   return {
