+++ packages/frontend/src/app/[postSlug]/page.tsx
@@ -1,16 +1,25 @@
+import { emitStructured } from '@kids-reporter/logger'
+import { headers } from 'next/headers'
 import { notFound, permanentRedirect } from 'next/navigation'
 
 import { getPostMeta } from '@/api/post'
-import { log, LogLevel } from '@/utils/log'
+import { getServerTraceHeaders } from '@/utils/trace-context'
 
 async function PostPage({ params }: { params: { postSlug: string } }) {
   const { postSlug } = params
-  const post = await getPostMeta({
-    where: { slug: postSlug },
-  })
+  const traceHeaders = getServerTraceHeaders(headers())
+  const post = await getPostMeta(
+    {
+      where: { slug: postSlug },
+    },
+    traceHeaders
+  )
 
   if (!post) {
-    log(LogLevel.WARNING, `Post not found! ${postSlug}`)
+    emitStructured({
+      severity: 'WARNING',
+      message: `Post not found! ${postSlug}`,
+    })
     notFound()
   }
