+++ packages/frontend/src/modules/membership/my-reading/post-question-answers-list/essay-answer-item.tsx
@@ -1,6 +1,7 @@
 'use client'
 
 import { PostEssayAnswer } from '__generated__/types'
+import { emitStructured } from '@kids-reporter/logger'
 import { useQueryClient } from '@tanstack/react-query'
 import errors from '@twreporter/errors'
 import { useState } from 'react'
@@ -18,7 +19,6 @@ import { getDisplayLikesCount } from '@/modules/idea-hub/utils'
 import { useHydratedAuthStore } from '@/services/auth/use-hydrated-auth-store'
 import BaodaozaiQAModal from '@/services/call-baodaozai/components/qa-modal'
 import { CallBaodaozaiProvider } from '@/services/call-baodaozai/context'
-import { log, LogLevel } from '@/utils'
 
 function EssayAnswerItem({
   answer,
@@ -80,7 +80,7 @@ function EssayAnswerItem({
         0
       )
 
-      log(LogLevel.ERROR, msg)
+      emitStructured({ severity: 'ERROR', message: msg })
       toast.error('編輯失敗，請稍後再試。')
     }
   }
