+++ packages/frontend/src/utils/index.ts
@@ -3,6 +3,5 @@ export * from './get-formatted-date'
 export * from './get-login-url'
 export * from './get-post-summaries'
 export * from './get-theme-color'
-export * from './log'
 export * from './media-query'
 export * from './send-rest-gql'
