+++ packages/frontend/src/utils/trace-context.ts
@@ -1,0 +1,23 @@
+import {
+  NormalizedTraceContext,
+  normalizeTraceContext,
+  type TraceHeaderInput,
+} from '@kids-reporter/logger'
+
+export function buildTraceHeaders(input?: TraceHeaderInput) {
+  const normalized = normalizeTraceContext(input, {
+    generateIfMissing: true,
+  })
+  return (
+    normalized?.traceHeaders ??
+    ({} as Partial<NormalizedTraceContext['traceHeaders']>)
+  )
+}
+
+export function getServerTraceHeaders(requestHeaders: Headers) {
+  const record: Record<string, string> = {}
+  requestHeaders.forEach((value, key) => {
+    record[key] = value
+  })
+  return buildTraceHeaders(record)
+}
